app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.121.3
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
requests==2.31.0
